Rewrote handling of token expiration to make it more robust and clean
