+++ services/web/modules/gitlab-sync/app/src/GitLabSyncController.mjs
@@ -116,8 +116,9 @@ async function oauth2Callback(req, res) {
 
   let token
   let refresh_token
+  let refresh_timestamp
   try {
-	[token, refresh_token] = await api.exchangeCodeForToken(code)
+	[token, refresh_token, refresh_timestamp] = await api.exchangeCodeForToken(code)
     if (!token || !refresh_token) {
       HttpErrorHandler.badRequest(req, res, 'Failed to obtain access token from Git server')
       return
@@ -132,7 +133,7 @@ async function oauth2Callback(req, res) {
   }
 
   try {
-    await TokenManager.saveUserToken(userId, {token, refresh_token})
+    await TokenManager.saveUserToken(userId, {token, refresh_token, refresh_timestamp})
   } catch (err) {
     const info = OError.getFullInfo(err)
     const errStatus  = info?.status || 500
