+++ services/web/modules/gitlab-sync/app/src/TokenManager.mjs
@@ -24,11 +24,61 @@ async function decryptAccessToken(tokenEncrypted) {
   }
 }
 
+async function refreshToken(userId, tokens) {
+	logger.info({ userId }, "refreshing GitLab token")
+
+	if (!tokens.refresh_token) {
+		logger.error("refresh_token invalid")
+		return false
+	}
+
+	let token
+	let refresh_token
+	let refresh_timestamp
+
+	try {
+		[token, refresh_token, refresh_timestamp] = await api.refreshToken(tokens.refresh_token)
+		if (!token || !refresh_token) {
+			HttpErrorHandler.badRequest(req, res, 'Failed to refresh access token')
+			return false
+		}
+	} catch (err) {
+		const info = OError.getFullInfo(err)
+		logger.error(OError.getFullStack(err))
+		logger.error({ info, userId }, 'Failed to refresh access token')
+		HttpErrorHandler.badRequest(req, res, err.message || 'Bad request')
+		return false
+	}
+
+	try {
+		await saveUserToken(userId, { token, refresh_token, refresh_timestamp })
+	} catch (err) {
+		const info = OError.getFullInfo(err)
+		const errStatus = info?.status || 500
+		logger.error(OError.getFullStack(err))
+		logger.error({ info, userId }, 'Error saving user token')
+		HttpErrorHandler.handleErrorByStatusCode(req, res, err, errStatus)
+		return false
+	}
+
+	return true
+}
+
 // ------------------------- exports -------------------------- //
 async function getUserToken(userId) {
   const credentials = await GitLabSyncUserCredentials.findOne(normalizeQuery({ userId }))
   if (!credentials) throw new InvalidTokenError('no user token', { userId, status: 400 })
-  return await decryptAccessToken(credentials.gitlab)
+  let tokens = await decryptAccessToken(credentials.gitlab)
+
+  const now = Math.floor(Date.now() / 1000);
+  if (!tokens.refresh_timestamp || tokens.refresh_timestamp < now) {
+	const refreshed = await refreshToken(userId, tokens)
+	if (refreshed) {
+		return await getUserToken(userId)
+	}
+  }
+
+  return tokens
 }
 
 async function saveUserToken(userId, accessToken) {
